+++ Mapper/changelog.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - AutoMapMarkers would spam error messages about unrevealed map for disabled object types.
 - Mapper wouldn't load on non-Windows systems because of the release zip file structure. ([#3](https://github.com/Genhis/VintageStoryMods/issues/3))
+- Mapper raised an exception when used in save files which had the world map disabled.
 - Compass didn't point to true north.
 
 ## [0.1.0] - 2026-01-07